% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_mapped}
\alias{example1_mapped}
\title{Differential gene expression results from \code{STRINGdb}}
\format{
A data frame of 17708 rows and 4 columns.
}
\source{
\url{https://www.bioconductor.org/packages/release/bioc/vignettes/STRINGdb/inst/doc/STRINGdb.pdf}
}
\usage{
example1_mapped
}
\description{
\code{diff_exp_example1} example dataset from \code{STRINGdb}, after mapping
gene names to the STRING database identifiers using the \code{map} method.
}
\keyword{datasets}
