% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cluster_of_int}
\alias{get_cluster_of_int}
\title{Get subnetwork of features of interest}
\usage{
get_cluster_of_int(all_clusters, hits_filt)
}
\arguments{
\item{all_clusters}{A list of clusters. Generated by \code{STRINGdb}'s method \code{get_clusters}.}

\item{hits_filt}{A data frame. A filtered data frame of features of interest of the mapped data frame generated by \code{STRINGdb}'s method \code{map}.}
}
\value{
Return a data frame with the information of which cluster the features are found and the number of features in those clusters.
}
\description{
\code{get_cluster_of_int} finds the clusters that the features of interest belong.
}
\examples{
\dontrun{
# get all clusters from STRING network using get_clusters method of STRINGdb
all_clusters <- string_db$get_clusters(hits)

get_cluster_of_int(all_clusters, hits_filt)
}
}
