% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_svg.R
\name{get_svg}
\alias{get_svg}
\title{Get a network SVG file from STRING website}
\usage{
get_svg(
  string_db,
  string_ids,
  required_score = NULL,
  network_flavor = c("evidence", "confidence", "actions"),
  file = NULL,
  payload_id = NULL
)
}
\arguments{
\item{string_db}{An instantiated STRINGdb reference class. See \code{\link[STRINGdb]{STRINGdb-class}}.}

\item{string_ids}{A vector of STRING IDs.}

\item{required_score}{An integer. The minimum STRING combined score of the interactions.}

\item{network_flavor}{One of "evidence", "confidence", "actions",
specify the flavor of the network}

\item{file}{Optional. Specify the file for the image.}

\item{payload_id}{Optional. See Payload Mechanism in \code{STRINGdb}'s \href{https://www.bioconductor.org/packages/release/bioc/vignettes/STRINGdb/inst/doc/STRINGdb.pdf}{vignette}.}
}
\value{
Return a XML document and a SVG file if \code{file} is specified.
}
\description{
\code{get_svg} is a modified version of \code{STRINGdb}'s \code{get_png}.
It returns a SVG image of a STRING network with the given identifiers.
}
\examples{
# save the XML document in a variable for downstream manipulation
xml <- get_svg(string_db, string_ids)

\dontrun{
# print a SVG file
get_svg(string_db, string_ids, file = "my_network.svg")
}

}
