% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.R
\name{plot_features}
\alias{plot_features}
\title{Get a network of features of interest from STRING website}
\usage{
plot_features(
  df_mapped,
  colors_vec,
  string_db,
  n_hits = 200,
  entire = FALSE,
  get_ann = FALSE
)
}
\arguments{
\item{df_mapped}{A data frame. Mapped data frame generated by \code{STRINGdb}'s method \code{map}.}

\item{colors_vec}{A named character vector of colors for feature of interest.}

\item{string_db}{An instantiated STRINGdb reference class. See \code{\link[STRINGdb]{STRINGdb-class}}.}

\item{n_hits}{Integer. The number of nodes to get from STRING network.}

\item{entire}{Logical. Return the entire STRING network if \code{TRUE}. Only clusters contain features of interest are returned if \code{FALSE}.}

\item{get_ann}{Logical. Add annotations to the table of features of interest.}
}
\value{
Return a SVG file of network of features of interest.
}
\description{
\code{plot_features} gets the STRING network or subnetwork of features of interests and save as a SVG file.
}
