% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string_db}
\alias{string_db}
\title{A \code{STRINGdb} reference class}
\format{
A \code{STRINGdb} reference class with 14 fields.
}
\source{
\url{https://www.bioconductor.org/packages/release/bioc/vignettes/STRINGdb/inst/doc/STRINGdb.pdf}
}
\usage{
string_db
}
\description{
The reference class is instantiated using STRING version 11, the species Human
and an interactions combined score threshold of 200 (see \code{STRINGdb}
vignette for more information).
}
\keyword{datasets}
