% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_nodes}
\alias{modify_nodes}
\title{Change color of features of interest and remove labels}
\usage{
modify_nodes(xml, colors_vec)
}
\arguments{
\item{xml}{A \code{xml_document} as defined in \code{xml2}.}

\item{colors_vec}{A named character vector of colors for feature of interest.}
}
\description{
\code{modify_nodes} modifies nodes of features of interest by assigning
user specified colors to these nodes and "grey" to the others and removing
labels from all nodes but those of interest. The idea and logic of this came
from the \code{change_STRING_colors.py} from STRING \href{https://string-db.org/}{website}.
}
\examples{
feature_of_int <- c("ATAD1", "BCL2L14", "MFGE8", "IDO1", "ABCA1")
colors_vec <- rep("rgb(101,226,11)", length(feature_of_int))
names(colors_vec) <- feature_of_int

}
